Override app with ws version in prod build
Fixes #405

Previously, the `--injectVersion` flag in the product build.js does not have any effect, because the version was not considered in the `generateMetaDataFile()` function.

This commit reads the workspace package.json and overrides the version of the default app with the version from the workspace package.json.

This behavior assumes that the product build.js file inserts the correct version number to the workspace package.json (i.e., without `--injectVersion` flag = default app version, with `--injectVersion` flag = version of product package.json).